MAGENTO2-209 Initial renaming of classes to prepare for structure change
