+++ composer.json
@@ -16,13 +16,14 @@
         "guzzlehttp/psr7": "^1.5"
     },
     "require-dev": {
+        "roave/security-advisories": "dev-master",
         "phpunit/phpunit": ">=5 <8",
         "php-coveralls/php-coveralls": "^2.1"
     },
     "autoload": {
         "psr-4": {
-            "Shopgate\\CloudIntegrationSdk\\": "src/",
-            "Shopgate\\CloudIntegrationSdk\\Tests\\": "tests/"
+            "Shopgate\\ConnectSdk\\": "src/",
+            "Shopgate\\ConnectSdk\\Tests\\": "tests/"
         }
     }
 }
